requirements.txt
numpy>=1.22.2
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0

psutil  # system utilization
py-cpuinfo  # display CPU info
